+++ utils/menu.sh
@@ -24,6 +24,7 @@ begins_with_short_option()
 # THE DEFAULTS INITIALIZATION - OPTIONALS
 _arg_full="off"
 _arg_local_backup="off"
+_arg_prefix="off"
 _arg_wordpress="off"
 
 dbsync_print_help() {
@@ -35,10 +36,11 @@ dbsync_print_help() {
 
   Make sure you are in a Magento 2 directory before trying to run any scripts.
 
-  Usage: db-sync.sh [-b|--local-backup] [-f|--full] [-w|--wordpress] [-v|--version] [-h|--help]
+  Usage: db-sync.sh [-b|--local-backup] [-f|--full] [-p|--prefix] [-w|--wordpress] [-v|--version] [-h|--help]
   Options:
     -b, --local-backup: perform a backup before importing remote (off by default)
     -f, --full: full database dump (off by default)
+    -p, --prefix, --no-prefix: prefix invoices, orders, shipments etc (off by default)
     -w, --wordpress: include wordpress content (off by default)"
     -v, --version: Prints version
     -h, --help: Prints help
@@ -75,6 +77,18 @@ parse_commandline()
 				if test -n "$_next" -a "$_next" != "$_key"
 				then
 					{ begins_with_short_option "$_next" && shift && set -- "-f" "-${_next}" "$@"; } || die "The short option '$_key' can't be decomposed to ${_key:0:2} and -${_key:2}, because ${_key:0:2} doesn't accept value and '-${_key:2:1}' doesn't correspond to a short option."
+				fi
+				;;
+			-p|--no-prefix|--prefix)
+				_arg_prefix="on"
+				test "${1:0:5}" = "--no-" && _arg_prefix="off"
+				;;
+			-p*)
+				_arg_prefix="on"
+				_next="${_key##-p}"
+				if test -n "$_next" -a "$_next" != "$_key"
+				then
+					{ begins_with_short_option "$_next" && shift && set -- "-p" "-${_next}" "$@"; } || die "The short option '$_key' can't be decomposed to ${_key:0:2} and -${_key:2}, because ${_key:0:2} doesn't accept value and '-${_key:2:1}' doesn't correspond to a short option."
 				fi
 				;;
 			-w|--no-wordpress|--wordpress)